Added workaround for bamboo collision post 1.21.80.

+++ src/main/java/ac/boar/anticheat/check/api/impl/OffsetHandlerCheck.java
@@ -8,6 +8,6 @@ public class OffsetHandlerCheck extends Check {
         super(player);
     }
 
-    public void onPredictionComplete(double offset) {
+    public void onPredictionComplete(float offset) {
     }
 }
