+++ src/main/java/ac/boar/anticheat/validator/BreakingBlockValidator.java
@@ -44,7 +44,7 @@ public final class BreakingBlockValidator {
                 return;
             }
 
-            final double distance = position.distance(player.position.toVector3i());
+            final float distance = position.distance(player.position.toVector3i());
             if (distance > 12) {
                 this.resyncBlock(position);
                 packet.setItemUseTransaction(null);
@@ -90,7 +90,7 @@ public final class BreakingBlockValidator {
             final BreakingData data = findCacheUsingPosition(action.getBlockPosition());
 
             if (data != null) {
-                final double distance = data.getPosition().distance(player.position.toVector3i());
+                final float distance = data.getPosition().distance(player.position.toVector3i());
                 if (distance > 12) {
                     this.resyncBlock(data.getPosition());
                     continue;
